feat: add layout with sidebar
