+++ src/pages/Jobs/All/index.tsx
@@ -1,7 +1,4 @@
-import { Box, Container, Grid, GridItem } from '@chakra-ui/react';
-
 import { Head } from '~/components/Head';
-import { Sidebar } from '~/components/Sidebar';
 
 import { ListJobs } from './ListJobs';
 
@@ -10,25 +7,7 @@ export const AllJobsPage = () => {
     <>
       <Head title="Meus Jobs" />
 
-      <Box as="section" w="100%">
-        <Grid
-          templateAreas={`"sidebar main"
-        "sidebar main"`}
-          gridTemplateColumns={'200px 1fr'}
-        >
-          <GridItem area={'sidebar'}>
-            <Box position="fixed" h="100%">
-              <Sidebar />
-            </Box>
-          </GridItem>
-
-          <GridItem area={'main'}>
-            <Container as="main" maxW="1440px" centerContent px="16">
-              <ListJobs />
-            </Container>
-          </GridItem>
-        </Grid>
-      </Box>
+      <ListJobs />
     </>
   );
 };
