+++ src/routes/index.tsx
@@ -3,6 +3,7 @@ import { Routes, Route } from 'react-router-dom';
 import { Loading } from '~/components/Loading';
 import { useSpinner } from '~/hooks/useSpinner';
 import { Auth } from '~/Layouts/Auth';
+import { WithSidebar } from '~/Layouts/WithSidebar';
 import { ForgotPassword } from '~/pages/Forgot';
 import { Home } from '~/pages/Home';
 import { AllJobsPage } from '~/pages/Jobs/All';
@@ -10,7 +11,7 @@ import { DetailsJobPage } from '~/pages/Jobs/Details';
 import { EditJobPage } from '~/pages/Jobs/Edit';
 import { EditJobReports } from '~/pages/Jobs/Edit/Reports';
 import { NewJobPage } from '~/pages/Jobs/New';
-import { ReportsPage } from '~/pages/Jobs/Reports';
+import { ReportsPage } from '~/pages/Reports';
 import { SignIn } from '~/pages/SignIn';
 import { SignUp } from '~/pages/Signup';
 import { ProtectedRoute } from '~/routes/ProtectedRoute';
@@ -29,13 +30,16 @@ export const GlobalRoutes = () => {
       </Route>
 
       <Route element={<ProtectedRoute />}>
-        <Route path="/dashboard" element={<Home />} />
+        <Route element={<WithSidebar />}>
+          <Route path="/dashboard" element={<Home />} />
+          <Route path="/jobs" element={<AllJobsPage />} />
+          <Route path="/jobs/reports" element={<ReportsPage />} />
+        </Route>
+
         <Route path="/jobs/:id" element={<DetailsJobPage />} />
         <Route path="/jobs/new" element={<NewJobPage />} />
         <Route path="/jobs/:id/edit" element={<EditJobPage />} />
         <Route path="/jobs/:id/cycles" element={<EditJobReports />} />
-        <Route path="/jobs" element={<AllJobsPage />} />
-        <Route path="/jobs/reports" element={<ReportsPage />} />
       </Route>
     </Routes>
   );
